[MOD/#360] NoteFragment와 NoteFrameFragment 분리
