+++ app/src/main/java/com/el/yello/presentation/main/profile/info/ProfileFragment.kt
@@ -101,15 +101,16 @@ class ProfileFragment : BindingFragment<FragmentProfileBinding>(R.layout.fragmen
     }
 
     private fun initAdapter() {
-        _adapter = ProfileFriendAdapter(viewModel,
+        _adapter = ProfileFriendAdapter(
+            viewModel,
             itemClick = { profileUserModel, position ->
                 initItemClickListener(
                     profileUserModel,
-                    position
+                    position,
                 )
             },
             shopClick = { initShopClickListener() },
-            modClick = { initProfileModClickListener() }
+            modClick = { initProfileModClickListener() },
         )
         binding.rvProfileFriendsList.adapter = adapter
     }
