+++ app/src/main/java/com/el/yello/presentation/main/yello/vote/VoteActivity.kt
@@ -6,7 +6,6 @@ import androidx.lifecycle.flowWithLifecycle
 import androidx.lifecycle.lifecycleScope
 import com.el.yello.R
 import com.el.yello.databinding.ActivityVoteBinding
-import com.el.yello.presentation.util.setCurrentItemWithDuration
 import com.el.yello.util.amplitude.AmplitudeUtils
 import com.el.yello.util.context.yelloSnackbar
 import com.example.ui.base.BindingActivity
@@ -24,9 +23,8 @@ class VoteActivity : BindingActivity<ActivityVoteBinding>(R.layout.activity_vote
 
     private val voteAdapter by lazy {
         VoteAdapter(
-            this,
-            viewModel.backgroundIndex,
-            viewModel.totalListCount + 1,
+            fragmentActivity = this,
+            voteListSize = viewModel.totalListCount + 1,
         )
     }
 
@@ -66,7 +64,6 @@ class VoteActivity : BindingActivity<ActivityVoteBinding>(R.layout.activity_vote
     private fun setupCurrentNoteIndex() {
         viewModel._currentNoteIndex.flowWithLifecycle(lifecycle)
             .onEach { index ->
-                binding.vpVote.setCurrentItemWithDuration(index, DURATION_NOTE_TRANSITION)
                 if (index <= viewModel.totalListCount) {
                     val properties = JSONObject().put(JSON_VOTE_STEP, index + 1)
                     AmplitudeUtils.trackEventWithProperties(EVENT_VIEW_VOTE_QUESTION, properties)
@@ -93,8 +90,6 @@ class VoteActivity : BindingActivity<ActivityVoteBinding>(R.layout.activity_vote
     }
 
     companion object {
-        private const val DURATION_NOTE_TRANSITION = 400L
-
         private const val JSON_VOTE_STEP = "vote_step"
         private const val EVENT_VIEW_VOTE_QUESTION = "view_vote_question"
     }
