+++ app/src/main/res/layout/activity_vote.xml
@@ -13,14 +13,25 @@
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/layout_vote"
-        android:background="@color/black"
         android:layout_width="match_parent"
         android:layout_height="match_parent">
 
+        <androidx.fragment.app.FragmentContainerView
+            android:id="@+id/fcv_vote"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintBottom_toBottomOf="parent"
+            tools:layout="@layout/fragment_note_frame"
+            android:layout_width="0dp"
+            android:layout_height="0dp" />
+
+        <!-- TODO : 동적으로 하위 버튼 만큼 위치 조정 -->
         <androidx.viewpager2.widget.ViewPager2
             android:id="@+id/vp_vote"
             android:layout_width="0dp"
-            android:layout_height="0dp"
+            android:layout_height="wrap_content"
+            tools:layout="@layout/fragment_note"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
